Consolidation of OnbaordingWorkflow
